COZMO-6361: Fix timing to report first event (#2832)
Change time counter to uint32 to avoid possible floating point round
errors.

Add check for “then == 0” to report first event that arrives in each
session.

+++ basestation/include/anki/cozmo/basestation/robotToEngineImplMessaging.h
@@ -93,7 +93,7 @@ private:
   double       _lastImageRecvTime  = -1.0;
   
   // For tracking time since last power level report (per accessory)
-  std::map<uint32_t, float> _lastPowerLevelSentTime;
+  std::map<uint32_t, uint32_t> _lastPowerLevelSentTime;
   
   // Internal helpers
   bool ShouldIgnoreMultipleImages() const;
